+++ cli/commands/migrate_squash.go
@@ -34,7 +34,10 @@ func newMigrateSquashCmd(ec *cli.ExecutionContext) *cobra.Command {
   hasura migrate squash --from 123
 
   # Add a name for the new squashed migration
-  hasura migrate squash --name "<name>" --from 123`,
+  hasura migrate squash --name "<name>" --from 123
+  
+  # squash all migrations from version 123 up to version 456:
+  hasura migrate squash --from 123 --to 456`,
 		SilenceUsage: true,
 		RunE: func(cmd *cobra.Command, args []string) error {
 			opts.newVersion = getTime()
@@ -48,6 +51,7 @@ func newMigrateSquashCmd(ec *cli.ExecutionContext) *cobra.Command {
 
 	f := migrateSquashCmd.Flags()
 	f.Uint64Var(&opts.from, "from", 0, "start squashing from this version")
+	f.Uint64Var(&opts.to, "to", 0, "squash until this version")
 	f.StringVar(&opts.name, "name", "squashed", "name for the new squashed migration")
 	f.BoolVar(&opts.deleteSource, "delete-source", false, "delete the source files after squashing without any confirmation")
 
@@ -61,6 +65,7 @@ type migrateSquashOptions struct {
 	EC *cli.ExecutionContext
 
 	from       uint64
+	to         uint64
 	name       string
 	newVersion int64
 
@@ -70,7 +75,14 @@ type migrateSquashOptions struct {
 
 func (o *migrateSquashOptions) run() error {
 	o.EC.Logger.Warnln("This command is currently experimental and hence in preview, correctness of squashed migration is not guaranteed!")
-	o.EC.Spin(fmt.Sprintf("Squashing migrations from %d to latest...", o.from))
+	if o.from > o.to && o.to != 0 {
+		return errors.Wrap(errors.New("incorrect version provided for to flag"), "the `to` flag has to have a version that is greater than that of `from` flag")
+	}
+	if o.to != 0 {
+		o.EC.Spin(fmt.Sprintf("Squashing migrations from %d to %d...", o.from, o.to))
+	} else {
+		o.EC.Spin(fmt.Sprintf("Squashing migrations from %d to latest...", o.from))
+	}
 	defer o.EC.Spinner.Stop()
 	migrateDrv, err := migrate.NewMigrate(o.EC, true, o.Database)
 	if err != nil {