+++ cli/migrate/api/squash.go
@@ -16,6 +16,7 @@ import (
 type squashCreateRequest struct {
 	Name    string `json:"name"`
 	From    uint64 `json:"from"`
+	To      uint64 `json:"to"`
 	version int64
 }
 
@@ -58,7 +59,11 @@ func SquashCreateAPI(c *gin.Context) {
 		c.JSON(http.StatusInternalServerError, &Response{Code: "internal_error", Message: err.Error()})
 		return
 	}
-	versions, err := cmd.SquashCmd(t, request.From, request.version, request.Name, sourceURL.Path)
+	if request.From > request.To && request.To != 0 {
+		c.JSON(http.StatusBadRequest, &Response{Code: "bad_request", Message: "`to` version has to be greater than `from` version"})
+		return
+	}
+	versions, err := cmd.SquashCmd(t, request.From, request.To, request.version, request.Name, sourceURL.Path)
 	if err != nil {
 		if strings.HasPrefix(err.Error(), DataAPIError) {
 			c.JSON(http.StatusBadRequest, &Response{Code: "data_api_error", Message: strings.TrimPrefix(err.Error(), DataAPIError)})